refactor hooks to eliminate useCollectionRefresh
